refactor: service CQRS 구조로 변경 (#110)
